refacto: exports interface motifs
